requirements.txt
python-telegram-bot
pyyaml
pysocks
pytils
sphinx
guzzle_sphinx_theme